+++ src/ecallisto_ng/__init__.py
@@ -15,4 +15,4 @@ working agreement.
 
 from __future__ import annotations
 
-__version__ = "0.7.1"
+__version__ = "0.7.2"